Add project structure and files
This commit addss the Features folder. This folder will contain
the .features files used to describe the feature to be tested and
defines the scenerio. Running the .feature files generates code that
can be used in the LoginSteps.java class that is located in the
step_definitions directory.

The LoginSteps class is composed of the code that was generated from
running the login.feature code.

TestRunner.java, so far, defines options for the test. Choose which
features are to be tested, and glue it. monocrhome option for making
the output cleaner. The plugin option is used for sending the output
from the test to a file (html,json) to a location where it can be opened
and viewed easily.