Add Parameterization and Data Driven Testing
This commit adds parameterization to the project by introducing
a datatable to the logindemo.feature file. This can be seen by adding
the <username> and <password> fields to the scenerio. Furthermore, a
datatable was added to the file as seen by the Examples: keyword followed
by a table created with the keywords and their values surrounded by pipes.

The LoginDemoSteps class was altered by changing the @When annotation
to use the reg ex expression for all (.*) when referring to the username
and password fields.This was done on the user_adds_username_and_password method.
The Strings for username and password are then passed into the method.