Add Background Demonstration
This commit adds te backgroundDemo.feature file, the BackgroundDemoSteps.java class
and the TestRunnerForBackgoundDemo.java class.

These files make up the framework for testing a project using a Background

A background is used when 2 or more scenerios share the same steps or
conditions when testing a feature. Example, a user will be logged in
to test the logout feature and navigation to a logged in user from the
same page.